refactor: critical issues 1.1 - naming strucuture aligned for components
