+++ docs/components/20250813-component-system-critical-issues-todo.md
@@ -1,0 +1,178 @@
+# Component System Critical Issues - Action Plan
+
+**Created:** 2025-08-13  
+**Priority:** High  
+**Scope:** Design System Scalability Improvements
+
+## Todo List
+
+### Phase 1: Foundation Standardization (High Priority)
+
+- [ ] **1.1** Audit all directory names for consistency
+  - [ ] Rename `accordionItems/` to `accordion-items/`
+  - [ ] Standardize all directories to kebab-case
+  - [ ] Update all import paths affected by renames
+  - [ ] Update barrel exports to reflect new structure
+
+- [ ] **1.2** Consolidate footer structure inconsistencies
+  - [ ] Remove empty `/footer/index.ts`
+  - [ ] Ensure all footer components are in `/layout/footer/`
+  - [ ] Update imports referencing footer components
+  - [ ] Verify no broken references exist
+
+- [ ] **1.3** Clean up incomplete directories
+  - [ ] Evaluate `/forms/` directory - populate or remove
+  - [ ] Add barrel exports to `/experimental/dialog/`
+  - [ ] Document experimental component status
+  - [ ] Create deprecation plan for unused directories
+
+### Phase 2: Design Token Implementation (High Priority)
+
+- [ ] **2.1** Create design token system
+  - [ ] Design spacing token scale (--spc-*)
+  - [ ] Design color token system (--color-*)
+  - [ ] Design typography token system
+  - [ ] Create token documentation
+
+- [ ] **2.2** Implement token usage
+  - [ ] Convert hardcoded values to tokens in primitives
+  - [ ] Update action components to use tokens
+  - [ ] Audit card components for token usage
+  - [ ] Create token validation utilities
+
+### Phase 3: Component API Standardization (Medium Priority)
+
+- [ ] **3.1** Document component interfaces
+  - [ ] Create TypeScript interfaces for all props
+  - [ ] Document event contracts
+  - [ ] Create usage examples for each component
+  - [ ] Add JSDoc comments to all exports
+
+- [ ] **3.2** Implement theming system
+  - [ ] Create variant prop patterns
+  - [ ] Implement theme context
+  - [ ] Add dark/light mode support
+  - [ ] Create theme switching utilities
+
+### Phase 4: Architecture Improvements (Medium Priority)
+
+- [ ] **4.1** Reduce directory complexity
+  - [ ] Flatten `/overlays/dialog/` structure if possible
+  - [ ] Consolidate similar components
+  - [ ] Remove duplicate functionality
+  - [ ] Optimize barrel export performance
+
+- [ ] **4.2** Add composition patterns
+  - [ ] Implement compound component patterns
+  - [ ] Add slot-based composition
+  - [ ] Create higher-order component utilities
+  - [ ] Document composition guidelines
+
+## Decision Points & Questions
+
+### 1. Directory Naming Strategy
+**Question:** Should we use kebab-case consistently or consider camelCase for nested contexts?  
+**Options:**
+- A) Strict kebab-case for all directories
+- B) kebab-case for top-level, camelCase for nested
+- C) Current mixed approach
+
+**Recommendation:** Option A - consistency trumps context
+
+### 2. Design Token Scope
+**Question:** How granular should the token system be?  
+**Considerations:**
+- Performance impact of CSS custom properties
+- Developer experience vs. flexibility
+- Migration effort from current CSS
+
+### 3. Component Deprecation Strategy
+**Question:** How should we handle experimental and deprecated components?  
+**Options:**
+- A) Remove immediately
+- B) Deprecate with timeline
+- C) Move to separate package
+
+### 4. Breaking Change Management
+**Question:** Should we batch breaking changes or implement incrementally?  
+**Impact:**
+- Development velocity
+- User experience
+- Testing complexity
+
+### 5. Theme System Architecture
+**Question:** Should theming be CSS-based or JavaScript-based?  
+**Considerations:**
+- Runtime performance
+- SSR compatibility
+- Developer experience
+
+## Success Criteria
+
+### Phase 1 Success Metrics
+- [ ] All directories follow consistent naming convention
+- [ ] Zero broken import references
+- [ ] All components have barrel exports
+- [ ] No empty or placeholder directories
+- [ ] Build passes without warnings
+
+### Phase 2 Success Metrics
+- [ ] Design token system covers 90% of current hardcoded values
+- [ ] All primitive components use tokens exclusively
+- [ ] Token documentation is complete and accessible
+- [ ] Performance regression < 5% on component rendering
+
+### Phase 3 Success Metrics
+- [ ] All components have TypeScript interfaces
+- [ ] 100% component documentation coverage
+- [ ] Theme switching works across all components
+- [ ] API consistency score > 90%
+
+### Phase 4 Success Metrics
+- [ ] Component discovery time reduced by 50%
+- [ ] Composition patterns documented with examples
+- [ ] Bundle size impact < 10% increase
+- [ ] Developer satisfaction survey shows improvement
+
+## Action Confidence Ratings
+
+### High Confidence (90-100%)
+- **1.1** Directory naming standardization - Well-defined scope, clear execution path
+- **1.2** Footer structure consolidation - Small scope, minimal dependencies
+- **2.1** Design token creation - Established patterns, clear benefits
+
+### Medium Confidence (70-89%)
+- **1.3** Directory cleanup - Requires stakeholder input on forms directory
+- **2.2** Token implementation - May reveal edge cases in existing components
+- **3.1** Component documentation - Time-intensive but straightforward
+
+### Lower Confidence (50-69%)
+- **3.2** Theming system - Complex architectural decisions required
+- **4.1** Directory flattening - May impact existing mental models
+- **4.2** Composition patterns - Requires design system expertise
+
+## Risk Assessment
+
+### High Risk
+- Breaking changes to import paths could impact entire application
+- Design token migration might reveal inconsistencies in current design
+
+### Medium Risk
+- Performance implications of token system not fully known
+- Theming system complexity could create maintenance burden
+
+### Low Risk
+- Documentation improvements have minimal implementation risk
+- Directory cleanup is largely mechanical
+
+## Next Steps
+
+1. **Immediate:** Begin Phase 1.1 directory naming audit
+2. **Week 1:** Complete directory standardization
+3. **Week 2:** Start design token system implementation
+4. **Week 3:** Begin component API documentation
+5. **Month 1 Review:** Assess progress and adjust timeline
+
+---
+
+**Note:** This plan should be reviewed and approved by the development team before implementation. Breaking changes require coordination with all stakeholders.
