+++ docs/components/20250813-issue-1.1-directory-naming-standardization.md
@@ -1,0 +1,164 @@
+# Issue 1.1: Directory Naming Standardization
+
+**Created:** 2025-08-13  
+**Priority:** High  
+**Confidence:** 90-100%  
+**Phase:** Foundation Standardization
+
+## Issue Description
+
+Inconsistent directory naming conventions across the component system create confusion and maintenance overhead. Currently using a mix of kebab-case and camelCase conventions.
+
+**Current State:**
+- `top-nav/` (kebab-case)
+- `accordionItems/` (camelCase)
+- Most directories use kebab-case
+- Inconsistency creates mental overhead
+
+## Todo List
+
+- [x] **Audit all directory names for consistency**
+  - [x] List all current directory names and their conventions
+  - [x] Identify directories that don't follow kebab-case
+  - [x] Document current import paths that will be affected
+
+- [x] **Rename non-conforming directories**
+  - [x] Rename `accordionItems/` to `accordion-items/`
+  - [x] Verify no other camelCase directories exist
+  - [x] Use git mv to preserve history
+
+- [x] **Update import paths**
+  - [x] Search for all imports referencing renamed directories
+  - [x] Update barrel export references
+  - [x] Update component imports in consuming files
+
+- [x] **Update barrel exports**
+  - [x] Modify index.ts files to reflect new structure
+  - [x] Verify all exports still function correctly
+  - [x] Test barrel export resolution
+
+- [x] **Verification**
+  - [x] Run build to ensure no broken imports
+  - [x] Test component functionality
+  - [x] Update documentation references
+
+## Decision Points & Questions
+
+### 1. Naming Convention Choice
+**Question:** Should we use kebab-case consistently for all directories?  
+**Options:**
+- A) Strict kebab-case for all directories ✅ **Recommended**
+- B) kebab-case for top-level, camelCase for nested
+- C) Current mixed approach
+
+**Rationale:** Consistency improves developer experience and reduces cognitive load.
+
+### 2. Migration Strategy
+**Question:** Should we rename directories incrementally or all at once?  
+**Considerations:**
+- Git history preservation
+- Team coordination
+- Testing complexity
+
+**Recommendation:** All at once to avoid prolonged inconsistency
+
+### 3. Documentation Updates
+**Question:** What documentation needs updating after renames?  
+**Scope:**
+- Component structure documentation
+- Import examples in README
+- Development guidelines
+
+## Success Criteria
+
+### Primary Success Metrics
+- [x] All directories follow kebab-case convention
+- [x] Zero broken import references
+- [x] Build passes without warnings
+- [x] All barrel exports function correctly
+- [x] Git history preserved for renamed directories
+
+### Secondary Success Metrics
+- [x] Documentation updated to reflect new structure
+- [ ] Developer team informed of changes
+- [x] IDE autocomplete works with new paths
+- [x] No performance regression in build times
+
+## Action Confidence Rating: 95%
+
+### High Confidence Factors
+- **Clear scope:** Limited to directory renames and import updates
+- **Mechanical changes:** Mostly find-and-replace operations
+- **Low risk:** Changes are structural, not functional
+- **Tooling support:** Git mv preserves history, IDEs help with refactoring
+
+### Potential Challenges
+- **Import discovery:** May miss some dynamic imports (5% risk)
+- **Third-party references:** External tools pointing to old paths (3% risk)
+- **Team coordination:** Ensuring all developers pull changes (2% risk)
+
+## Risk Assessment
+
+### High Risk
+- None identified for this specific change
+
+### Medium Risk
+- **Missed import references:** Some imports might be in config files or tests
+- **External dependencies:** Third-party tools or documentation
+
+### Low Risk
+- **Build failures:** Easy to detect and fix
+- **Developer confusion:** Temporary, resolved with communication
+
+## Implementation Plan
+
+### Phase 1: Audit (Day 1)
+1. List all current directory names
+2. Identify non-kebab-case directories
+3. Search codebase for affected import paths
+4. Document findings
+
+### Phase 2: Rename (Day 1-2)
+1. Use `git mv` to rename directories
+2. Update barrel exports immediately
+3. Update import paths in batches
+4. Test after each batch
+
+### Phase 3: Verification (Day 2)
+1. Run full build and test suite
+2. Verify all components render correctly
+3. Update documentation
+4. Communicate changes to team
+
+### Rollback Plan
+- Keep backup branch before changes
+- Git history allows easy reversion
+- Import path changes are easily reversible
+
+## Related Issues
+
+- **Issue 1.2:** Footer structure consolidation (may create additional directories)
+- **Issue 1.3:** Directory cleanup (may remove some directories)
+
+## Implementation Results
+
+### Completed Actions
+1. **Directory Audit:** Identified only `accordionItems/` as non-kebab-case directory
+2. **Rename Executed:** Successfully renamed to `accordion-items/` using `git mv`
+3. **Import Analysis:** No direct imports found requiring updates
+4. **Build Verification:** `npm run build` completed successfully with no errors
+5. **Type Check:** `npm run check` passed with 0 errors and 0 warnings
+6. **Documentation:** Updated `FRONT_END_PROJECT_REFERENCE.md`
+
+### Time to Complete
+**Actual Effort:** 1 hour (significantly less than estimated 4-6 hours)
+
+### Breaking Changes
+**Actual Impact:** None - no direct imports to the renamed directory were found
+
+---
+
+**Status:** ✅ **COMPLETED**  
+**Date Completed:** 2025-08-13  
+**Breaking Changes:** None (originally anticipated but not required)  
+**Team Coordination Required:** Minimal - no active imports affected
