+++ docs/components/20250813-issue-1.2-footer-structure-consolidation.md
@@ -1,0 +1,169 @@
+# Issue 1.2: Footer Structure Consolidation
+
+**Created:** 2025-08-13  
+**Priority:** High  
+**Confidence:** 90-100%  
+**Phase:** Foundation Standardization
+
+## Issue Description
+
+Inconsistent footer component organization with empty placeholder directory and actual components in different location creates confusion about component location and structure.
+
+**Current State:**
+- Empty `/footer/index.ts` directory
+- Actual footer components in `/layout/footer/`
+- Inconsistent mental model for footer location
+- Potential for duplicate footer implementations
+
+## Todo List
+
+- [ ] **Audit footer component structure**
+  - [ ] Verify `/footer/index.ts` is empty/placeholder
+  - [ ] Catalog all footer components in `/layout/footer/`
+  - [ ] Check for any imports referencing empty footer directory
+  - [ ] Document current footer component relationships
+
+- [ ] **Remove empty footer directory**
+  - [ ] Delete `/footer/index.ts`
+  - [ ] Remove `/footer/` directory
+  - [ ] Verify no git submodules or special files
+
+- [ ] **Verify footer component consolidation**
+  - [ ] Confirm all footer components are in `/layout/footer/`
+  - [ ] Check footer component barrel exports
+  - [ ] Verify footer components are properly exported from layout
+
+- [ ] **Update import references**
+  - [ ] Search for imports referencing `/footer/`
+  - [ ] Update any imports to use `/layout/footer/` path
+  - [ ] Update barrel export references if needed
+
+- [ ] **Documentation cleanup**
+  - [ ] Update component structure documentation
+  - [ ] Remove footer from top-level component list
+  - [ ] Ensure layout documentation includes footer
+
+## Decision Points & Questions
+
+### 1. Footer Location Strategy
+**Question:** Should footer components remain in layout or move to top-level?  
+**Options:**
+- A) Keep in `/layout/footer/` ✅ **Recommended**
+- B) Move to top-level `/footer/`
+- C) Split between layout and standalone
+
+**Rationale:** Footer is inherently a layout component and belongs with other layout components.
+
+### 2. Footer Component Scope
+**Question:** What constitutes a footer component vs. layout component?  
+**Considerations:**
+- Footer-specific components (FooterTitle, Footer)
+- Reusable components used in footer
+- Layout utilities used by footer
+
+**Recommendation:** Keep footer-specific components in `/layout/footer/`
+
+### 3. Import Path Consistency
+**Question:** Should layout components be imported via layout barrel or direct?  
+**Current:** `import { Footer } from '$lib/components/layout';`  
+**Alternative:** `import { Footer } from '$lib/components/layout/footer';`
+
+**Recommendation:** Maintain current layout barrel pattern
+
+## Success Criteria
+
+### Primary Success Metrics
+- [ ] No empty or placeholder directories exist
+- [ ] All footer components are in consistent location
+- [ ] Zero broken import references
+- [ ] Footer components accessible via layout barrel
+- [ ] Build passes without warnings
+
+### Secondary Success Metrics
+- [ ] Component structure documentation updated
+- [ ] No confusion about footer component location
+- [ ] Consistent mental model for layout components
+- [ ] No duplicate footer functionality
+
+## Action Confidence Rating: 95%
+
+### High Confidence Factors
+- **Small scope:** Only removing empty directory and updating imports
+- **Clear structure:** Footer components already properly located
+- **Low complexity:** Minimal dependencies to update
+- **Easy verification:** Simple to test and validate
+
+### Potential Challenges
+- **Hidden references:** Some imports might be in config or test files (3% risk)
+- **Documentation lag:** Forgetting to update related documentation (2% risk)
+
+## Risk Assessment
+
+### High Risk
+- None identified for this specific change
+
+### Medium Risk
+- **Missed import cleanup:** Some references to empty footer directory might remain
+
+### Low Risk
+- **Build failures:** Easy to detect and fix
+- **Documentation inconsistency:** Non-functional but creates confusion
+
+## Implementation Plan
+
+### Phase 1: Audit (30 minutes)
+1. Verify `/footer/index.ts` contents
+2. List all components in `/layout/footer/`
+3. Search for imports referencing `/footer/`
+4. Document current structure
+
+### Phase 2: Cleanup (30 minutes)
+1. Remove empty `/footer/` directory
+2. Update any import references found
+3. Verify layout barrel exports include footer
+4. Test build
+
+### Phase 3: Verification (15 minutes)
+1. Run build and ensure no errors
+2. Verify footer components render correctly
+3. Update component structure documentation
+4. Confirm no broken references
+
+### Rollback Plan
+- Simple to recreate empty directory if needed
+- Git history preserves deleted files
+- No functional changes to rollback
+
+## Related Issues
+
+- **Issue 1.1:** Directory naming standardization (affects layout directory structure)
+- **Issue 1.3:** Directory cleanup (part of broader cleanup effort)
+
+## Files Affected
+
+### To Remove
+- `/frontend/src/lib/components/footer/index.ts`
+- `/frontend/src/lib/components/footer/` (directory)
+
+### To Verify
+- `/frontend/src/lib/components/layout/footer/Footer.svelte`
+- `/frontend/src/lib/components/layout/footer/FooterTitle.svelte`
+- `/frontend/src/lib/components/layout/footer/index.ts`
+- `/frontend/src/lib/components/layout/index.ts`
+
+### To Update
+- Any imports referencing `/footer/`
+- Component structure documentation
+
+## Next Steps
+
+1. **Immediate:** Audit footer directory structure
+2. **Today:** Remove empty directory and cleanup imports
+3. **Today:** Update documentation
+4. **Monitor:** Watch for any issues in subsequent builds
+
+---
+
+**Estimated Effort:** 1-2 hours  
+**Breaking Changes:** Minimal - only if imports reference empty directory  
+**Team Coordination Required:** Low - internal cleanup
