+++ docs/components/20250813-issue-1.3-directory-cleanup.md
@@ -1,0 +1,207 @@
+# Issue 1.3: Directory Cleanup
+
+**Created:** 2025-08-13  
+**Priority:** High  
+**Confidence:** 70-89%  
+**Phase:** Foundation Standardization
+
+## Issue Description
+
+Several directories are incomplete, empty, or contain experimental components without proper organization, creating uncertainty about component status and cluttering the component structure.
+
+**Current State:**
+- Empty `/forms/` directory
+- `/experimental/dialog/` components lack barrel exports
+- Unclear status of experimental components
+- No deprecation strategy for unused directories
+
+## Todo List
+
+- [ ] **Audit incomplete directories**
+  - [ ] Evaluate `/forms/` directory purpose and contents
+  - [ ] Document experimental component status and usage
+  - [ ] Identify any other empty or placeholder directories
+  - [ ] Check for orphaned or unused components
+
+- [ ] **Forms directory decision**
+  - [ ] Determine if forms functionality is needed
+  - [ ] Either populate with form components or remove entirely
+  - [ ] Document decision rationale
+  - [ ] Update component roadmap if removing
+
+- [ ] **Experimental component organization**
+  - [ ] Add barrel exports to `/experimental/dialog/`
+  - [ ] Document experimental component purpose
+  - [ ] Create experimental component guidelines
+  - [ ] Establish promotion/deprecation criteria
+
+- [ ] **Create deprecation strategy**
+  - [ ] Define component lifecycle stages
+  - [ ] Create deprecation warning system
+  - [ ] Establish timeline for experimental component decisions
+  - [ ] Document experimental-to-stable promotion process
+
+- [ ] **Directory structure cleanup**
+  - [ ] Remove truly unused directories
+  - [ ] Add README files for experimental areas
+  - [ ] Update component inventory documentation
+  - [ ] Verify all directories have clear purpose
+
+## Decision Points & Questions
+
+### 1. Forms Directory Strategy
+**Question:** Should we populate or remove the forms directory?  
+**Options:**
+- A) Remove entirely - forms not needed in current scope
+- B) Populate with basic form components
+- C) Keep as placeholder for future development ✅ **Requires stakeholder input**
+
+**Considerations:**
+- Current application form requirements
+- Future development plans
+- Component library completeness
+
+### 2. Experimental Component Strategy
+**Question:** How should we handle experimental components?  
+**Options:**
+- A) Remove immediately if not used
+- B) Deprecate with timeline ✅ **Recommended**
+- C) Move to separate package
+- D) Promote to stable if functional
+
+**Rationale:** Deprecation allows for evaluation period and prevents accidental removal of valuable work.
+
+### 3. Component Lifecycle Management
+**Question:** What stages should components go through?  
+**Proposed stages:**
+- Experimental → Alpha → Beta → Stable → Deprecated → Removed
+
+**Timeline:** 
+- Experimental: No timeline limit
+- Alpha: 2-4 weeks evaluation
+- Beta: 4-8 weeks testing
+- Deprecated: 8-12 weeks notice
+
+### 4. Barrel Export Strategy for Experimental
+**Question:** Should experimental components have barrel exports?  
+**Considerations:**
+- Makes imports cleaner
+- Suggests stability/permanence
+- Easier to track usage
+
+**Recommendation:** Yes, but with clear experimental labeling
+
+## Success Criteria
+
+### Primary Success Metrics
+- [ ] All directories have clear purpose and contents
+- [ ] No empty or placeholder directories without justification
+- [ ] Experimental components have proper organization
+- [ ] Component lifecycle strategy documented
+- [ ] All components have barrel exports where appropriate
+
+### Secondary Success Metrics
+- [ ] Developer confusion about component status eliminated
+- [ ] Component discovery improved
+- [ ] Future component development process clarified
+- [ ] Documentation updated to reflect clean structure
+
+## Action Confidence Rating: 75%
+
+### Medium Confidence Factors
+- **Requires stakeholder input:** Forms directory decision needs product/design input
+- **Subjective decisions:** What constitutes "experimental" vs "alpha"
+- **Future planning:** Decisions affect long-term component strategy
+- **Usage analysis needed:** Must determine if experimental components are used
+
+### High Confidence Factors
+- **Clear technical steps:** Adding barrel exports is straightforward
+- **Documentation improvements:** Always beneficial
+- **Structure cleanup:** Obvious improvements available
+
+### Uncertainty Factors
+- **Forms requirements:** Unknown if forms functionality needed (20% uncertainty)
+- **Experimental usage:** Don't know if dialog components are actively used (15% uncertainty)
+- **Stakeholder priorities:** Component library scope decisions (10% uncertainty)
+
+## Risk Assessment
+
+### High Risk
+- **Premature removal:** Deleting components that are actually needed
+- **Breaking changes:** Removing imports that are used elsewhere
+
+### Medium Risk
+- **Decision paralysis:** Stakeholder input delays may stall progress
+- **Scope creep:** Directory cleanup may reveal larger architectural issues
+
+### Low Risk
+- **Documentation updates:** Non-functional improvements
+- **Barrel export additions:** Purely additive changes
+
+## Implementation Plan
+
+### Phase 1: Research & Analysis (Day 1)
+1. Audit all directory contents and purposes
+2. Search codebase for usage of experimental components
+3. Identify stakeholders for forms directory decision
+4. Document current component lifecycle (if any)
+
+### Phase 2: Quick Wins (Day 1-2)
+1. Add barrel exports to experimental components
+2. Add README files for experimental directories
+3. Remove clearly unused directories (after verification)
+4. Update component inventory
+
+### Phase 3: Stakeholder Decisions (Day 3-5)
+1. Present forms directory options to stakeholders
+2. Get decision on experimental component strategy
+3. Implement decisions based on stakeholder input
+4. Update documentation with new policies
+
+### Phase 4: Implementation (Day 5-7)
+1. Execute stakeholder decisions
+2. Implement component lifecycle documentation
+3. Clean up remaining directory issues
+4. Verify all changes work correctly
+
+### Rollback Plan
+- Git history preserves any deleted components
+- Barrel exports are purely additive
+- Documentation changes are easily reversible
+- Stakeholder decisions can be revised
+
+## Related Issues
+
+- **Issue 1.1:** Directory naming (may affect experimental directory names)
+- **Issue 1.2:** Footer consolidation (part of broader cleanup)
+- **Issue 3.1:** Component documentation (will help clarify component status)
+
+## Stakeholder Input Required
+
+### Forms Directory Decision
+**Stakeholders:** Product Manager, Design Lead, Frontend Lead  
+**Questions:**
+- Are form components needed in current product scope?
+- What form functionality is required?
+- Timeline for form component development?
+
+### Experimental Component Strategy
+**Stakeholders:** Frontend Lead, Architecture Team  
+**Questions:**
+- How should experimental components be managed?
+- What criteria determine promotion to stable?
+- Should experimental components be user-facing?
+
+## Next Steps
+
+1. **Immediate:** Begin directory audit and usage analysis
+2. **Day 1:** Add barrel exports to experimental components
+3. **Day 2:** Schedule stakeholder meetings for decisions
+4. **Week 1:** Implement stakeholder decisions
+5. **Week 2:** Monitor for any issues or missed components
+
+---
+
+**Estimated Effort:** 8-12 hours (including stakeholder coordination)  
+**Breaking Changes:** Potential - depends on stakeholder decisions  
+**Team Coordination Required:** High - requires stakeholder input
