+++ frontend/src/routes/+layout.svelte
@@ -2,7 +2,7 @@
   import '../app.css';
 
   import { Footer } from '$lib/components/footer';
-  import TopNav from '$lib/components/top-nav/TopNav.svelte';
+  import { TopNav } from '$lib/components/top-nav';
 
   import { onMount } from 'svelte';
   import { beforeNavigate } from '$app/navigation';
